Albert-3D-stokes (#9)
Expand the visco-elastic Stokes solver to 3D problems. The main modification to the library is the division of the visco-elastic kernels and solvers into two submodules, one for 2D and the other one for 3D. This requires some adjustments in MetaJustRelax.jl. 

Add benchmarks for 3D viscous and visco-elastic Stokes, and corresponding tests.

Reworked MetaStokes.jl to accommodate 3D structs